+++ forte2/sci/sci_helper_sigma_build.cc
@@ -3,40 +3,12 @@
 #include "helpers/ndarray.h"
 #include "helpers/np_vector_functions.h"
 #include "helpers/np_matrix_functions.h"
+#include "helpers/parallel.h"
 #include "helpers/spin.h"
-
-#include <future>
 
 #include "sci_helper.h"
 
 namespace forte2 {
-
-namespace {
-
-// A helper function to run a loop in parallel over a given number of indices. The work function is
-// called with the index as argument.
-template <typename WorkFn>
-void run_parallel_indices(size_t count, size_t num_threads, WorkFn&& work) {
-    if (num_threads <= 1 || count == 0) {
-        for (size_t i{0}; i < count; ++i)
-            work(i);
-        return;
-    }
-
-    std::vector<std::future<void>> workers;
-    workers.reserve(num_threads);
-    for (size_t t{0}; t < num_threads; ++t) {
-        workers.push_back(std::async(std::launch::async, [count, num_threads, t, &work]() {
-            for (size_t i{t}; i < count; i += num_threads)
-                work(i);
-        }));
-    }
-
-    for (auto& w : workers)
-        w.get();
-}
-
-} // namespace
 
 void SelectedCIHelper::compute_det_energies() {
     // compute the energy of all the determinants
@@ -234,372 +206,6 @@ void SelectedCIHelper::H2ab(std::span<double> basis, std::span<double> sigma) co
             }
         }
     });
-}
-
-double SelectedCIHelper::find_matching_dets_1rdm(size_t left_root, size_t right_root,
-                                                 const SelectedCIStrings& list, size_t i, size_t j,
-                                                 double sign) const {
-    double result = 0.0;
-
-    // Find the range of determinants with the current alpha string
-    const auto& [istart, iend] = list.range(i);
-    const auto& [jstart, jend] = list.range(j);
-    const auto& det_permutation = list.det_permutation();
-
-    // Here we choose to loop over the smaller range and look up the determinants in the larger
-    // range by using the hash map
-    if (iend - istart >= jend - jstart) {
-        const auto& i_map = list.second_string_to_det_index()[i];
-        for (size_t jj{jstart}; jj < jend; ++jj) {
-            const auto idx_j = list.sorted_dets_second_string(jj);
-            if (const auto it = i_map.find(idx_j); it != i_map.end()) {
-                result += sign * c_[nroots_ * it->second + left_root] *
-                          c_[nroots_ * det_permutation[jj] + right_root];
-            }
-        }
-    } else {
-        const auto& j_map = list.second_string_to_det_index()[j];
-        for (size_t ii{istart}; ii < iend; ++ii) {
-            const auto idx_i = list.sorted_dets_second_string(ii);
-            if (const auto it = j_map.find(idx_i); it != j_map.end()) {
-                result += sign * c_[nroots_ * det_permutation[ii] + left_root] *
-                          c_[nroots_ * it->second + right_root];
-            }
-        }
-    }
-
-    return result;
-}
-
-np_matrix SelectedCIHelper::compute_a_1rdm(size_t left_root, size_t right_root) const {
-    auto rdm = make_zeros<nb::numpy, double, 2>({norb_, norb_});
-    double* rdm_data = rdm.data();
-
-    const auto first_string_size = ab_list_.first_string_size();
-    const auto& one_hole_first_strings = ab_list_.one_hole_first_strings();
-    // Loop over all unique alpha strings
-    for (size_t i{0}; i < first_string_size; ++i) {
-        const auto& sublist = ab_list_.one_hole_first_string_list()[i];
-        for (const auto& [p, hole_idx, sign_p] : sublist) {
-            const auto& inv_sublist = ab_list_.one_hole_first_string_list_inv()[hole_idx];
-            for (const auto& [q, j, sign_q] : inv_sublist) {
-                const double sign = sign_p * sign_q;
-                rdm_data[p * norb_ + q] +=
-                    find_matching_dets_1rdm(left_root, right_root, ab_list_, i, j, sign);
-            }
-        }
-    }
-
-    return rdm;
-}
-
-np_matrix SelectedCIHelper::compute_b_1rdm(size_t left_root, size_t right_root) const {
-    auto rdm = make_zeros<nb::numpy, double, 2>({norb_, norb_});
-    double* rdm_data = rdm.data();
-
-    const auto first_string_size = ba_list_.first_string_size();
-    const auto& one_hole_first_strings = ba_list_.one_hole_first_strings();
-    // Loop over all unique beta strings
-    for (size_t i{0}; i < first_string_size; ++i) {
-        const auto& sublist = ba_list_.one_hole_first_string_list()[i];
-        for (const auto& [p, hole_idx, sign_p] : sublist) {
-            const auto& inv_sublist = ba_list_.one_hole_first_string_list_inv()[hole_idx];
-            for (const auto& [q, j, sign_q] : inv_sublist) {
-                const double sign = sign_p * sign_q;
-                rdm_data[p * norb_ + q] +=
-                    find_matching_dets_1rdm(left_root, right_root, ba_list_, i, j, sign);
-            }
-        }
-    }
-
-    return rdm;
-}
-
-np_matrix SelectedCIHelper::compute_sf_1rdm(size_t left_root, size_t right_root) const {
-    auto sf_1rdm = make_zeros<nb::numpy, double, 2>({norb_, norb_});
-    if (norb_ > 0) {
-        auto a_1rdm = compute_a_1rdm(left_root, right_root);
-        auto b_1rdm = compute_b_1rdm(left_root, right_root);
-        matrix::daxpy<double>(1.0, a_1rdm, sf_1rdm);
-        matrix::daxpy<double>(1.0, b_1rdm, sf_1rdm);
-    }
-    return sf_1rdm;
-}
-
-np_matrix SelectedCIHelper::compute_aa_2rdm(size_t left_root, size_t right_root) const {
-    // calculate the number of pairs of orbitals p > q
-    const size_t npairs = (norb_ * (norb_ - 1)) / 2;
-    auto rdm = make_zeros<nb::numpy, double, 2>({npairs, npairs});
-    double* rdm_data = rdm.data();
-
-    const auto first_string_size = ab_list_.first_string_size();
-    const auto& one_hole_first_strings = ab_list_.one_hole_first_strings();
-    // Loop over all unique alpha strings
-    for (size_t i{0}; i < first_string_size; ++i) {
-        const auto& sublist = ab_list_.two_hole_string_list()[i];
-        for (const auto& [p, q, hole_idx, sign_pq] : sublist) { // (p < q)
-            const size_t pq = pair_index_gt(p, q);
-            const auto& inv_sublist = ab_list_.two_hole_string_list_inv()[hole_idx];
-            for (const auto& [r, s, j, sign_rs] : inv_sublist) { // (r < s)
-                const size_t rs = pair_index_gt(r, s);
-                const double sign = sign_pq * sign_rs;
-                rdm_data[pq * npairs + rs] +=
-                    find_matching_dets_1rdm(left_root, right_root, ab_list_, i, j, sign);
-            }
-        }
-    }
-    return rdm;
-}
-
-np_matrix SelectedCIHelper::compute_bb_2rdm(size_t left_root, size_t right_root) const {
-    // calculate the number of pairs of orbitals p > q
-    const size_t npairs = (norb_ * (norb_ - 1)) / 2;
-    auto rdm = make_zeros<nb::numpy, double, 2>({npairs, npairs});
-    double* rdm_data = rdm.data();
-
-    const auto first_string_size = ba_list_.first_string_size();
-    const auto& one_hole_first_strings = ba_list_.one_hole_first_strings();
-    // Loop over all unique alpha strings
-    for (size_t i{0}; i < first_string_size; ++i) {
-        const auto& sublist = ba_list_.two_hole_string_list()[i];
-        for (const auto& [p, q, hole_idx, sign_pq] : sublist) { // (p < q)
-            const size_t pq = pair_index_gt(p, q);
-            const auto& inv_sublist = ba_list_.two_hole_string_list_inv()[hole_idx];
-            for (const auto& [r, s, j, sign_rs] : inv_sublist) { // (r < s)
-                const size_t rs = pair_index_gt(r, s);
-                const double sign = sign_pq * sign_rs;
-                rdm_data[pq * npairs + rs] +=
-                    find_matching_dets_1rdm(left_root, right_root, ba_list_, i, j, sign);
-            }
-        }
-    }
-    return rdm;
-}
-
-np_tensor4 SelectedCIHelper::compute_ab_2rdm(size_t left_root, size_t right_root) const {
-    auto rdm = make_zeros<nb::numpy, double, 4>({norb_, norb_, norb_, norb_});
-    double* rdm_data = rdm.data();
-
-    const auto first_string_size = ab_list_.first_string_size();
-    const auto& det_permutation = ab_list_.det_permutation();
-    // Loop over all unique alpha strings
-    for (size_t i{0}; i < first_string_size; ++i) {
-        const auto& i_map = ab_list_.second_string_to_det_index()[i];
-        // Loop over all single excitations in the alpha string.
-        // a+_q a_p |i_a> -> +/-|j_a>
-        const auto& sublist_a = ab_list_.one_hole_first_string_list()[i];
-        for (const auto& [p, hole_idx, sign_p] : sublist_a) {
-            const auto& inv_sublist_a = ab_list_.one_hole_first_string_list_inv()[hole_idx];
-            for (const auto& [q, j, sign_q] : inv_sublist_a) {
-                const auto& [jstart, jend] = ab_list_.range(j);
-                // Loop over all the beta strings with the same alpha string
-                for (size_t jj{jstart}; jj < jend; ++jj) {
-                    const auto idx_j = ab_list_.sorted_dets_second_string(jj);
-                    // Loop over single excitations in the beta string.
-                    // a+_s a_r |j_b> -> +/-|k_b>
-                    const auto& sublist_b = ab_list_.one_hole_second_string_list()[idx_j];
-                    for (const auto& [r, hole_idx_b, sign_r] : sublist_b) {
-                        const auto& inv_sublist_b =
-                            ab_list_.one_hole_second_string_list_inv()[hole_idx_b];
-                        for (const auto& [s, k, sign_s] : inv_sublist_b) {
-                            const double sign = sign_p * sign_q * sign_r * sign_s;
-                            // Here we get (leaving note since this is a bit subtle):
-                            //     <j_a k_b|a+_r a_s a+_q a_p|i_a i_b>
-                            //   = <j_a k_b|a+_q a+_r a_s a_p|i_a i_b> = gamma2(q_a,r_b,p_a,s_b)
-                            // Check if the determinant with the new beta string exists
-                            if (const auto it = i_map.find(k); it != i_map.end()) {
-                                rdm_data[q * norb3_ + r * norb2_ + p * norb_ + s] +=
-                                    sign * c_[nroots_ * it->second + left_root] *
-                                    c_[nroots_ * det_permutation[jj] + right_root];
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    return rdm;
-}
-
-np_tensor4 SelectedCIHelper::compute_sf_2rdm(size_t left_root, size_t right_root) const {
-    auto rdm_sf = make_zeros<nb::numpy, double, 4>({norb_, norb_, norb_, norb_});
-
-    if (norb_ < 1) {
-        return rdm_sf; // No 2-RDM for less than 1 orbitals
-    }
-
-    auto rdm_sf_v = rdm_sf.view();
-    // Mixed-spin contribution (1 orbital or more)
-    {
-        auto rdm_ab = compute_ab_2rdm(left_root, right_root);
-        auto rdm_ab_v = rdm_ab.view();
-        for (size_t p{0}; p < norb_; ++p) {
-            for (size_t q{0}; q < norb_; ++q) {
-                for (size_t r{0}; r < norb_; ++r) {
-                    for (size_t s{0}; s < norb_; ++s) {
-                        rdm_sf_v(p, q, r, s) += rdm_ab_v(p, q, r, s) + rdm_ab_v(q, p, s, r);
-                    }
-                }
-            }
-        }
-    }
-
-    if (norb_ < 2) {
-        return rdm_sf; // No same-spin contributions to the 2-RDM for less than 2 orbitals
-    }
-
-    // To reduce the  memory footprint, we compute the aa and bb contributions in a packed
-    // format and one at a time.
-    for (auto spin : {Spin::Alpha, Spin::Beta}) {
-        auto rdm_ss = spin == Spin::Alpha ? compute_aa_2rdm(left_root, right_root)
-                                          : compute_bb_2rdm(left_root, right_root);
-        auto rdm_ss_v = rdm_ss.view();
-        for (size_t p{1}, pq{0}; p < norb_; ++p) {
-            for (size_t q{0}; q < p; ++q, ++pq) { // p > q
-                for (size_t r{1}, rs{0}; r < norb_; ++r) {
-                    for (size_t s{0}; s < r; ++s, ++rs) { // r > s
-                        auto element = rdm_ss_v(pq, rs);
-                        rdm_sf_v(p, q, r, s) += element;
-                        rdm_sf_v(q, p, r, s) -= element;
-                        rdm_sf_v(p, q, s, r) -= element;
-                        rdm_sf_v(q, p, s, r) += element;
-                    }
-                }
-            }
-        }
-    }
-
-    return rdm_sf;
-}
-
-double SelectedCIHelper::find_matching_dets_1trdm(size_t left_root, size_t right_root,
-                                                  const SelectedCIStrings& left_list,
-                                                  const SelectedCIStrings& right_list,
-                                                  const std::vector<double>& left_c,
-                                                  const std::vector<double>& right_c,
-                                                  size_t left_nroots, size_t right_nroots, size_t i,
-                                                  size_t j, double sign) const {
-    double result = 0.0;
-
-    // Find the range of determinants with the current alpha string
-    const auto& [istart, iend] = left_list.range(i);
-    const auto& [jstart, jend] = right_list.range(j);
-    const auto& right_det_permutation = right_list.det_permutation();
-    const auto& left_det_permutation = left_list.det_permutation();
-
-    // Here we choose to loop over the smaller range and look up the determinants in the larger
-    // range by using the hash map. The map keys are list-local second-string indices, so we
-    // map from the right <-> left actual second-string index.
-    if (iend - istart >= jend - jstart) {
-        const auto& i_map = left_list.second_string_to_det_index()[i];
-        for (size_t jj{jstart}; jj < jend; ++jj) {
-            const auto right_idx_j = right_list.sorted_dets_second_string(jj);
-            // find the corresponding second string in the left list and get its index
-            const auto& second_string_j = right_list.sorted_second_string(right_idx_j);
-            const auto left_idx_j = left_list.find_second_string_index(second_string_j);
-            if (left_idx_j.has_value()) {
-                if (const auto it = i_map.find(*left_idx_j); it != i_map.end()) {
-                    result += sign * left_c[left_nroots * it->second + left_root] *
-                              right_c[right_nroots * right_det_permutation[jj] + right_root];
-                }
-            }
-        }
-    } else {
-        const auto& j_map = right_list.second_string_to_det_index()[j];
-        for (size_t ii{istart}; ii < iend; ++ii) {
-            const auto left_idx_i = left_list.sorted_dets_second_string(ii);
-            // find the corresponding second string in the right list and get its index
-            const auto& second_string_i = left_list.sorted_second_string(left_idx_i);
-            const auto right_idx_i = right_list.find_second_string_index(second_string_i);
-            if (right_idx_i.has_value()) {
-                if (const auto it = j_map.find(*right_idx_i); it != j_map.end()) {
-                    result += sign * left_c[left_nroots * left_det_permutation[ii] + left_root] *
-                              right_c[right_nroots * it->second + right_root];
-                }
-            }
-        }
-    }
-
-    return result;
-}
-
-np_matrix SelectedCIHelper::compute_s_1trdm(const SelectedCIHelper& right_helper, size_t left_root,
-                                            size_t right_root, Spin spin) const {
-    const auto& left_helper = *this;
-
-    if (left_helper.norb_ != right_helper.norb_) {
-        throw std::runtime_error("SCI transition RDMs: The number of MOs must be the same in the "
-                                 "two wave functions.");
-    }
-    if (left_helper.na_ != right_helper.na_ || left_helper.nb_ != right_helper.nb_) {
-        throw std::runtime_error("SCI transition RDMs: The number of alpha and beta electrons must "
-                                 "be the same in the two wave functions.");
-    }
-    if (left_root >= left_helper.nroots_ || right_root >= right_helper.nroots_) {
-        throw std::runtime_error("SCI transition RDMs: Root index out of range.");
-    }
-
-    const auto& left_c = left_helper.c_;
-    const auto& right_c = right_helper.c_;
-    const auto left_nroots = left_helper.nroots_;
-    const auto right_nroots = right_helper.nroots_;
-
-    auto rdm = make_zeros<nb::numpy, double, 2>({norb_, norb_});
-    double* rdm_data = rdm.data();
-
-    // pick the appropriate string lists based on the spin
-    const auto& left_list = is_alpha(spin) ? left_helper.ab_list_ : left_helper.ba_list_;
-    const auto& right_list = is_alpha(spin) ? right_helper.ab_list_ : right_helper.ba_list_;
-
-    const auto right_first_string_size = right_list.first_string_size();
-    const auto& right_one_hole_first_strings = right_list.one_hole_first_strings();
-    const auto& left_one_hole_first_strings_index = left_list.one_hole_first_strings_index();
-
-    // Loop over all unique strings of the right state
-    for (size_t j{0}; j < right_first_string_size; ++j) {
-        const auto& sublist_right = right_list.one_hole_first_string_list()[j];
-        // loop over all single excitations in the right string. a_p |j> -> +/-|k>
-        for (const auto& [p, right_hole_idx, sign_p] : sublist_right) {
-            // get the one-hole alpha string from the right solver
-            const auto& K = right_one_hole_first_strings[right_hole_idx];
-            // find the index of the one-hole alpha string K in the left solver
-            if (const auto it = left_one_hole_first_strings_index.find(K);
-                it != left_one_hole_first_strings_index.end()) {
-                // if found, get the corresponding inv_sublist on the left
-                const auto& inv_sublist = left_list.one_hole_first_string_list_inv()[it->second];
-                for (const auto& [q, i, sign_q] : inv_sublist) {
-                    const double sign = sign_p * sign_q;
-                    rdm_data[q * norb_ + p] += find_matching_dets_1trdm(
-                        left_root, right_root, left_list, right_list, left_c, right_c, left_nroots,
-                        right_nroots, i, j, sign);
-                }
-            }
-        }
-    }
-
-    return rdm;
-}
-
-np_matrix SelectedCIHelper::compute_a_1trdm(const SelectedCIHelper& right_helper, size_t left_root,
-                                            size_t right_root) const {
-    return compute_s_1trdm(right_helper, left_root, right_root, Spin::Alpha);
-}
-
-np_matrix SelectedCIHelper::compute_b_1trdm(const SelectedCIHelper& right_helper, size_t left_root,
-                                            size_t right_root) const {
-    return compute_s_1trdm(right_helper, left_root, right_root, Spin::Beta);
-}
-
-np_matrix SelectedCIHelper::compute_sf_1trdm(const SelectedCIHelper& right_helper, size_t left_root,
-                                             size_t right_root) const {
-    auto sf_1trdm = make_zeros<nb::numpy, double, 2>({norb_, norb_});
-    if (norb_ > 0) {
-        auto a_1trdm = compute_a_1trdm(right_helper, left_root, right_root);
-        auto b_1trdm = compute_b_1trdm(right_helper, left_root, right_root);
-        matrix::daxpy<double>(1.0, a_1trdm, sf_1trdm);
-        matrix::daxpy<double>(1.0, b_1trdm, sf_1trdm);
-    }
-    return sf_1trdm;
 }
 
 } // namespace forte2
